requirements.txt
django<1.12
djangorestframework<3.8
psycopg2-binary<2.8
python-dateutil<2.8
requests<2.2
gunicorn==19.6.0
raven==6.1.0
dj-database-url==0.3.0
whitenoise==3.2